Refactor:Validator 모아놓기
