+++ src/main/java/lotto/model/Validator.java
@@ -1,0 +1,32 @@
+package lotto.model;
+
+import java.util.List;
+
+public class Validator {
+    public static int validatePurchaseAmount(String input) {
+        try {
+            int amount = Integer.parseInt(input);
+            if (amount % 1000 != 0) {
+                throw new IllegalArgumentException("[ERROR] 구입 금액은 1,000원 단위여야 합니다.");
+            }
+            return amount;
+        } catch (NumberFormatException e) {
+            throw new IllegalArgumentException("[ERROR] 올바른 숫자를 입력해 주세요.");
+        }
+    }
+
+    public static int validateBonusNumber(String input, List<Integer> winningNumbers) {
+        try {
+            int bonusNumber = Integer.parseInt(input);
+            if (bonusNumber < 1 || bonusNumber > 45) {
+                throw new IllegalArgumentException("[ERROR] 보너스 번호는 1부터 45 사이의 숫자여야 합니다.");
+            }
+            if (winningNumbers.contains(bonusNumber)) {
+                throw new IllegalArgumentException("[ERROR] 보너스 번호는 당첨 번호와 중복될 수 없습니다.");
+            }
+            return bonusNumber;
+        } catch (NumberFormatException e) {
+            throw new IllegalArgumentException("[ERROR] 보너스 번호는 숫자여야 합니다.");
+        }
+    }
+}
