+++ src/main/java/lotto/view/InputView.java
@@ -2,6 +2,7 @@ package lotto.view;
 
 import camp.nextstep.edu.missionutils.Console;
 import lotto.model.Lotto;
+import lotto.model.Validator;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -10,7 +11,7 @@ public class InputView {
     public static int getPurchaseAmount() {
         System.out.println("구입금액을 입력해 주세요.");
         String input = Console.readLine().trim();
-        return validatePurchaseAmount(input);
+        return Validator.validatePurchaseAmount(input);
     }
 
     public static List<Integer> getWinningNumbers() {
@@ -22,19 +23,7 @@ public class InputView {
     public static int getBonusNumber(List<Integer> winningNumbers) {
         System.out.println("보너스 번호를 입력해 주세요.");
         String input = Console.readLine().trim();
-        return validateBonusNumber(input, winningNumbers);
-    }
-
-    public static int validatePurchaseAmount(String input) {
-        try {
-            int amount = Integer.parseInt(input);
-            if (amount % 1000 != 0) {
-                throw new IllegalArgumentException("[ERROR] 구입 금액은 1,000원 단위여야 합니다.");
-            }
-            return amount;
-        } catch (NumberFormatException e) {
-            throw new IllegalArgumentException("[ERROR] 올바른 숫자를 입력해 주세요.");
-        }
+        return Validator.validateBonusNumber(input, winningNumbers);
     }
 
     private static List<Integer> parseNumbers(String input) {
@@ -52,19 +41,5 @@ public class InputView {
         return numbers;
     }
 
-    public static int validateBonusNumber(String input, List<Integer> winningNumbers) {
-        try {
-            int bonusNumber = Integer.parseInt(input);
-            if (bonusNumber < 1 || bonusNumber > 45) {
-                throw new IllegalArgumentException("[ERROR] 보너스 번호는 1부터 45 사이의 숫자여야 합니다.");
-            }
-            if (winningNumbers.contains(bonusNumber)) {
-                throw new IllegalArgumentException("[ERROR] 보너스 번호는 당첨 번호와 중복될 수 없습니다.");
-            }
-            return bonusNumber;
-        } catch (NumberFormatException e) {
-            throw new IllegalArgumentException("[ERROR] 보너스 번호는 숫자여야 합니다.");
-        }
-    }
 }
