+++ src/test/java/lotto/ApplicationTest.java
@@ -1,6 +1,7 @@
 package lotto;
 
 import camp.nextstep.edu.missionutils.test.NsTest;
+import lotto.model.Validator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -60,14 +61,14 @@ class ApplicationTest extends NsTest {
     @DisplayName("입력한 금액이 1000원 단위가 아니면 예외가 발생한다.")
     @Test
     void 입력한_금액이_1000단위가_아니면_예외가_발생한다() {
-        assertThatThrownBy(() -> InputView.validatePurchaseAmount("1500"))
+        assertThatThrownBy(() -> Validator.validatePurchaseAmount("1500"))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
     @DisplayName("입력한 금액이 숫자가 아니면 예외가 발생한다.")
     @Test
     void 입력한_금액이_숫자가_아니면_예외가_발생한다() {
-        assertThatThrownBy(() -> InputView.validatePurchaseAmount("abc"))
+        assertThatThrownBy(() -> Validator.validatePurchaseAmount("abc"))
                 .isInstanceOf(IllegalArgumentException.class);
     }
