+++ src/test/java/lotto/LottoTest.java
@@ -2,6 +2,7 @@ package lotto;
 
 import lotto.model.Lotto;
 import lotto.view.InputView;
+import lotto.model.Validator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -36,24 +37,24 @@ class LottoTest {
     @DisplayName("보너스 번호가 1~45 범위를 벗어나면 예외가 발생한다.")
     @Test
     void 보너스_번호가_범위를_벗어나면_예외가_발생한다() {
-        assertThatThrownBy(() -> InputView.validateBonusNumber("0", List.of(1, 2, 3, 4, 5, 6)))
+        assertThatThrownBy(() -> Validator.validateBonusNumber("0", List.of(1, 2, 3, 4, 5, 6)))
                 .isInstanceOf(IllegalArgumentException.class);
 
-        assertThatThrownBy(() -> InputView.validateBonusNumber("46", List.of(1, 2, 3, 4, 5, 6)))
+        assertThatThrownBy(() -> Validator.validateBonusNumber("46", List.of(1, 2, 3, 4, 5, 6)))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
     @DisplayName("보너스 번호가 당첨 번호와 중복되면 예외가 발생한다.")
     @Test
     void 보너스_번호가_당첨_번호와_중복되면_예외가_발생한다() {
-        assertThatThrownBy(() -> InputView.validateBonusNumber("5", List.of(1, 2, 3, 4, 5, 6)))
+        assertThatThrownBy(() -> Validator.validateBonusNumber("5", List.of(1, 2, 3, 4, 5, 6)))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
     @DisplayName("보너스 번호가 숫자가 아니면 예외가 발생한다.")
     @Test
     void 보너스_번호가_숫자가_아니면_예외가_발생한다() {
-        assertThatThrownBy(() -> InputView.validateBonusNumber("five", List.of(1, 2, 3, 4, 5, 6)))
+        assertThatThrownBy(() -> Validator.validateBonusNumber("five", List.of(1, 2, 3, 4, 5, 6)))
                 .isInstanceOf(IllegalArgumentException.class);
     }
     // TODO: 추가 기능 구현에 따른 테스트 코드 작성
